backend/requirements.txt
torch==2.2.2+cpu
numpy==1.26.4
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-pptx==0.6.23
sentence-transformers==3.2.1
faiss-cpu==1.8.0
requests==2.32.3
pydantic==2.9.2
pydantic-settings==2.6.1
feedparser==6.0.11
